fix: various linting issues, removing old linting exclusions
