+++ packages/parser/src/validators/spec/openapi.ts
@@ -2,142 +2,40 @@ import type { OpenAPIV3, OpenAPIV3_1 } from 'openapi-types';
 
 import { ono } from '@jsdevtools/ono';
 
-import { supportedHTTPMethods, pathParameterTemplateRegExp } from '../../util.js';
+import { supportedHTTPMethods, pathParameterTemplateRegExp } from '../../lib/index.js';
 
-type OpenAPIDocument = OpenAPIV3_1.Document | OpenAPIV3.Document;
 type ParameterObject =
   | (OpenAPIV3_1.ParameterObject | OpenAPIV3_1.ReferenceObject)
   | (OpenAPIV3.ParameterObject | OpenAPIV3.ReferenceObject);
 
 /**
- * Validates parts of the OpenAPI 3.0 and 3.1 that aren't covered by their JSON Schema definitions.
+ * Checks the given parameter list for duplicates.
  *
- * @todo This library currently does not validate required properties like the Swagger validator
- * does due to some gnarly quirks with cases where a required property exists within a `oneOf` or
- * `anyOf` (and within a child `allOf` of one of those). See
- * https://api.apis.guru/v2/specs/twitter.com/labs/2.13/openapi.yaml for a good example.
- *
- * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.0.md}
- * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md}
  */
-export function validateSpec(api: OpenAPIDocument) {
-  const operationIds: string[] = [];
-  Object.keys(api.paths || {}).forEach(pathName => {
-    const path = api.paths[pathName];
-    const pathId = `/paths${pathName}`;
+function checkForDuplicates(params: ParameterObject[]) {
+  for (let i = 0; i < params.length - 1; i++) {
+    const outer = params[i];
+    for (let j = i + 1; j < params.length; j++) {
+      const inner = params[j];
+      if ('$ref' in outer || '$ref' in inner) {
+        continue;
+      }
 
-    if (path && pathName.startsWith('/')) {
-      validatePath(path, pathId, operationIds);
-    }
-  });
-
-  /**
-   * There's a problem with how the 3.0 schema uses `patternProperties` for defining the format of
-   * scheme names that it ignores anything that doesn't match, so if you for example have a space
-   * in a schema name it'll be seen as valid when it should instead trigger a validation error.
-   *
-   * @see {@link https://github.com/APIDevTools/swagger-parser/issues/184}
-   */
-  if ('openapi' in api && api.openapi.startsWith('3.0')) {
-    if (api.components) {
-      Object.keys(api.components).forEach((componentType: keyof typeof api.components) => {
-        Object.keys(api.components[componentType]).forEach(componentName => {
-          const componentId = `/components/${componentType}/${componentName}`;
-
-          if (!/^[a-zA-Z0-9.\-_]+$/.test(componentName)) {
-            throw ono.syntax(
-              `Validation failed. ${componentId} has an invalid name. Component names should match against: /^[a-zA-Z0-9.-_]+$/`,
-            );
-          }
-        });
-      });
+      if (outer.name === inner.name && outer.in === inner.in) {
+        throw ono.syntax(`Validation failed. Found multiple ${outer.in} parameters named "${outer.name}"`);
+      }
     }
   }
 }
 
 /**
- * Validates the given path.
+ * Validates the given Swagger schema object.
  *
  */
-function validatePath(
-  path: OpenAPIV3_1.PathItemObject | OpenAPIV3.PathItemObject,
-  pathId: string,
-  operationIds: string[],
-) {
-  supportedHTTPMethods.forEach(operationName => {
-    const operation = path[operationName];
-    const operationId = `${pathId}/${operationName}`;
-
-    if (operation) {
-      const declaredOperationId = operation.operationId;
-      if (declaredOperationId) {
-        if (!operationIds.includes(declaredOperationId)) {
-          operationIds.push(declaredOperationId);
-        } else {
-          throw ono.syntax(`Validation failed. Duplicate operation id '${declaredOperationId}'`);
-        }
-      }
-
-      validateParameters(path, pathId, operation, operationId);
-
-      Object.keys(operation.responses || {}).forEach(responseCode => {
-        const response = operation.responses[responseCode];
-        const responseId = `${operationId}/responses/${responseCode}`;
-        if (response && !('$ref' in response)) {
-          validateResponse(response, responseId);
-        }
-      });
-    }
-  });
-}
-
-/**
- * Validates the parameters for the given operation.
- *
- */
-function validateParameters(
-  path: OpenAPIV3_1.PathItemObject | OpenAPIV3.PathItemObject,
-  pathId: string,
-  operation: OpenAPIV3_1.OperationObject | OpenAPIV3.OperationObject,
-  operationId: string,
-) {
-  const pathParams = path.parameters || [];
-  const operationParams = operation.parameters || [];
-
-  // Check for duplicate path parameters.
-  try {
-    checkForDuplicates(pathParams);
-  } catch (e) {
-    throw ono.syntax(e, `Validation failed. ${pathId} has duplicate parameters`);
+function validateSchema(schema: OpenAPIV3_1.SchemaObject | OpenAPIV3.SchemaObject, schemaId: string) {
+  if (schema.type === 'array' && !schema.items) {
+    throw ono.syntax(`Validation failed. ${schemaId} is an array, so it must include an "items" schema`);
   }
-
-  // Check for duplicate operation parameters.
-  try {
-    checkForDuplicates(operationParams);
-  } catch (e) {
-    throw ono.syntax(e, `Validation failed. ${operationId} has duplicate parameters`);
-  }
-
-  // Combine the path and operation parameters, with the operation params taking precedence over
-  // the path params.
-  const params = pathParams.reduce((combinedParams, value) => {
-    const duplicate = combinedParams.some(param => {
-      if ('$ref' in param || '$ref' in value) {
-        return false;
-      }
-
-      return param.in === value.in && param.name === value.name;
-    });
-
-    if (!duplicate) {
-      combinedParams.push(value);
-    }
-
-    return combinedParams;
-  }, operationParams.slice());
-
-  validatePathParameters(params, pathId, operationId);
-  validateParameterTypes(params, operationId);
 }
 
 /**
@@ -205,23 +103,52 @@ function validateParameterTypes(params: ParameterObject[], operationId: string) 
 }
 
 /**
- * Checks the given parameter list for duplicates.
+ * Validates the parameters for the given operation.
  *
  */
-function checkForDuplicates(params: ParameterObject[]) {
-  for (let i = 0; i < params.length - 1; i++) {
-    const outer = params[i];
-    for (let j = i + 1; j < params.length; j++) {
-      const inner = params[j];
-      if ('$ref' in outer || '$ref' in inner) {
-        continue;
+function validateParameters(
+  path: OpenAPIV3_1.PathItemObject | OpenAPIV3.PathItemObject,
+  pathId: string,
+  operation: OpenAPIV3_1.OperationObject | OpenAPIV3.OperationObject,
+  operationId: string,
+) {
+  const pathParams = path.parameters || [];
+  const operationParams = operation.parameters || [];
+
+  // Check for duplicate path parameters.
+  try {
+    checkForDuplicates(pathParams);
+  } catch (e) {
+    throw ono.syntax(e, `Validation failed. ${pathId} has duplicate parameters`);
+  }
+
+  // Check for duplicate operation parameters.
+  try {
+    checkForDuplicates(operationParams);
+  } catch (e) {
+    throw ono.syntax(e, `Validation failed. ${operationId} has duplicate parameters`);
+  }
+
+  // Combine the path and operation parameters, with the operation params taking precedence over
+  // the path params.
+  const params = pathParams.reduce((combinedParams, value) => {
+    const duplicate = combinedParams.some(param => {
+      if ('$ref' in param || '$ref' in value) {
+        return false;
       }
 
-      if (outer.name === inner.name && outer.in === inner.in) {
-        throw ono.syntax(`Validation failed. Found multiple ${outer.in} parameters named "${outer.name}"`);
-      }
+      return param.in === value.in && param.name === value.name;
+    });
+
+    if (!duplicate) {
+      combinedParams.push(value);
     }
-  }
+
+    return combinedParams;
+  }, operationParams.slice());
+
+  validatePathParameters(params, pathId, operationId);
+  validateParameterTypes(params, operationId);
 }
 
 /**
@@ -263,11 +190,83 @@ function validateResponse(response: OpenAPIV3_1.ResponseObject | OpenAPIV3.Respo
 }
 
 /**
- * Validates the given Swagger schema object.
+ * Validates the given path.
  *
  */
-function validateSchema(schema: OpenAPIV3_1.SchemaObject | OpenAPIV3.SchemaObject, schemaId: string) {
-  if (schema.type === 'array' && !schema.items) {
-    throw ono.syntax(`Validation failed. ${schemaId} is an array, so it must include an "items" schema`);
+function validatePath(
+  path: OpenAPIV3_1.PathItemObject | OpenAPIV3.PathItemObject,
+  pathId: string,
+  operationIds: string[],
+) {
+  supportedHTTPMethods.forEach(operationName => {
+    const operation = path[operationName];
+    const operationId = `${pathId}/${operationName}`;
+
+    if (operation) {
+      const declaredOperationId = operation.operationId;
+      if (declaredOperationId) {
+        if (!operationIds.includes(declaredOperationId)) {
+          operationIds.push(declaredOperationId);
+        } else {
+          throw ono.syntax(`Validation failed. Duplicate operation id '${declaredOperationId}'`);
+        }
+      }
+
+      validateParameters(path, pathId, operation, operationId);
+
+      Object.keys(operation.responses || {}).forEach(responseCode => {
+        const response = operation.responses[responseCode];
+        const responseId = `${operationId}/responses/${responseCode}`;
+        if (response && !('$ref' in response)) {
+          validateResponse(response, responseId);
+        }
+      });
+    }
+  });
+}
+
+/**
+ * Validates parts of the OpenAPI 3.0 and 3.1 that aren't covered by their JSON Schema definitions.
+ *
+ * @todo This library currently does not validate required properties like the Swagger validator
+ * does due to some gnarly quirks with cases where a required property exists within a `oneOf` or
+ * `anyOf` (and within a child `allOf` of one of those). See
+ * https://api.apis.guru/v2/specs/twitter.com/labs/2.13/openapi.yaml for a good example.
+ *
+ * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.0.md}
+ * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md}
+ */
+export function validateSpec(api: OpenAPIV3_1.Document | OpenAPIV3.Document) {
+  const operationIds: string[] = [];
+  Object.keys(api.paths || {}).forEach(pathName => {
+    const path = api.paths[pathName];
+    const pathId = `/paths${pathName}`;
+
+    if (path && pathName.startsWith('/')) {
+      validatePath(path, pathId, operationIds);
+    }
+  });
+
+  /**
+   * There's a problem with how the 3.0 schema uses `patternProperties` for defining the format of
+   * scheme names that it ignores anything that doesn't match, so if you for example have a space
+   * in a schema name it'll be seen as valid when it should instead trigger a validation error.
+   *
+   * @see {@link https://github.com/APIDevTools/swagger-parser/issues/184}
+   */
+  if ('openapi' in api && api.openapi.startsWith('3.0')) {
+    if (api.components) {
+      Object.keys(api.components).forEach((componentType: keyof typeof api.components) => {
+        Object.keys(api.components[componentType]).forEach(componentName => {
+          const componentId = `/components/${componentType}/${componentName}`;
+
+          if (!/^[a-zA-Z0-9.\-_]+$/.test(componentName)) {
+            throw ono.syntax(
+              `Validation failed. ${componentId} has an invalid name. Component names should match against: /^[a-zA-Z0-9.-_]+$/`,
+            );
+          }
+        });
+      });
+    }
   }
 }
