+++ packages/parser/src/validators/spec/swagger.ts
@@ -2,120 +2,69 @@ import type { IJsonSchema, OpenAPIV2 } from 'openapi-types';
 
 import { ono } from '@jsdevtools/ono';
 
-import { swaggerHTTPMethods, pathParameterTemplateRegExp } from '../../util.js';
+import { swaggerHTTPMethods, pathParameterTemplateRegExp } from '../../lib/index.js';
 
 /**
- * Validates parts of the Swagger 2.0 specification that aren't covered by its JSON Schema
- * definition.
+ * Checks the given parameter list for duplicates.
  *
- * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/2.0.md}
  */
-export function validateSpec(api: OpenAPIV2.Document) {
-  const operationIds: string[] = [];
-  Object.keys(api.paths || {}).forEach(pathName => {
-    const path = api.paths[pathName];
-    const pathId = `/paths${pathName}`;
-
-    if (path && pathName.startsWith('/')) {
-      validatePath(api, path, pathId, operationIds);
+function checkForDuplicates(params: OpenAPIV2.ParameterObject[]) {
+  for (let i = 0; i < params.length - 1; i++) {
+    const outer = params[i];
+    for (let j = i + 1; j < params.length; j++) {
+      const inner = params[j];
+      if (outer.name === inner.name && outer.in === inner.in) {
+        throw ono.syntax(`Validation failed. Found multiple ${outer.in} parameters named "${outer.name}"`);
+      }
     }
-  });
-
-  Object.keys(api.definitions || {}).forEach(definitionName => {
-    const definition = api.definitions[definitionName];
-    const definitionId = `/definitions/${definitionName}`;
-
-    if (!/^[a-zA-Z0-9.\-_]+$/.test(definitionName)) {
-      throw ono.syntax(
-        `Validation failed. ${definitionId} has an invalid name. Definition names should match against: /^[a-zA-Z0-9.-_]+$/`,
-      );
-    }
-
-    validateRequiredPropertiesExist(definition, definitionId);
-  });
+  }
 }
 
 /**
- * Validates the given path.
+ * Validates that the declared properties of the given Swagger schema object actually exist.
  *
  */
-function validatePath(api: OpenAPIV2.Document, path: OpenAPIV2.PathItemObject, pathId: string, operationIds: string[]) {
-  swaggerHTTPMethods.forEach(operationName => {
-    const operation = path[operationName];
-    const operationId = `${pathId}/${operationName}`;
-
-    if (operation) {
-      const declaredOperationId = operation.operationId;
-      if (declaredOperationId) {
-        if (!operationIds.includes(declaredOperationId)) {
-          operationIds.push(declaredOperationId);
-        } else {
-          throw ono.syntax(`Validation failed. Duplicate operation id '${declaredOperationId}'`);
+function validateRequiredPropertiesExist(schema: IJsonSchema, schemaId: string) {
+  // Recursively collects all properties of the schema and its ancestors. They are added to the props object.
+  function collectProperties(schemaObj: IJsonSchema, props: Record<string, IJsonSchema>) {
+    if (schemaObj.properties) {
+      Object.keys(schemaObj.properties).forEach(property => {
+        // eslint-disable-next-line no-prototype-builtins
+        if (schemaObj.properties.hasOwnProperty(property)) {
+          // eslint-disable-next-line no-param-reassign
+          props[property] = schemaObj.properties[property];
         }
-      }
-      validateParameters(api, path, pathId, operation, operationId);
-
-      Object.keys(operation.responses || {}).forEach(responseName => {
-        const response = operation.responses[responseName];
-        if ('$ref' in response || !response) {
-          return;
-        }
-
-        const responseId = `${operationId}/responses/${responseName}`;
-        validateResponse(responseName, response, responseId);
       });
     }
-  });
+
+    if (schemaObj.allOf) {
+      schemaObj.allOf.forEach(parent => {
+        collectProperties(parent, props);
+      });
+    }
+  }
+
+  if (schema.required && Array.isArray(schema.required)) {
+    const props: Record<string, IJsonSchema> = {};
+    collectProperties(schema, props);
+    schema.required.forEach(requiredProperty => {
+      if (!props[requiredProperty]) {
+        throw ono.syntax(
+          `Validation failed. Property '${requiredProperty}' listed as required but does not exist in '${schemaId}'`,
+        );
+      }
+    });
+  }
 }
 
 /**
- * Validates the parameters for the given operation.
+ * Validates the given Swagger schema object.
  *
  */
-function validateParameters(
-  api: OpenAPIV2.Document,
-  path: OpenAPIV2.PathItemObject,
-  pathId: string,
-  operation: OpenAPIV2.OperationObject,
-  operationId: string,
-) {
-  const pathParams = (path.parameters || []).filter(param => !('$ref' in param)) as OpenAPIV2.ParameterObject[];
-  const operationParams = (operation.parameters || []).filter(
-    param => !('$ref' in param),
-  ) as OpenAPIV2.ParameterObject[];
-
-  // Check for duplicate path parameters
-  try {
-    checkForDuplicates(pathParams);
-  } catch (e) {
-    throw ono.syntax(e, `Validation failed. ${pathId} has duplicate parameters`);
+function validateSchema(schema: OpenAPIV2.SchemaObject, schemaId: string) {
+  if (schema.type === 'array' && !schema.items) {
+    throw ono.syntax(`Validation failed. ${schemaId} is an array, so it must include an "items" schema`);
   }
-
-  // Check for duplicate operation parameters
-  try {
-    checkForDuplicates(operationParams);
-  } catch (e) {
-    throw ono.syntax(e, `Validation failed. ${operationId} has duplicate parameters`);
-  }
-
-  // Combine the path and operation parameters,
-  // with the operation params taking precedence over the path params
-  const params = pathParams.reduce((combinedParams, value) => {
-    const duplicate = combinedParams.some(param => {
-      if ('$ref' in param || '$ref' in value) {
-        return false;
-      }
-      return param.in === value.in && param.name === value.name;
-    });
-    if (!duplicate) {
-      combinedParams.push(value);
-    }
-    return combinedParams;
-  }, operationParams.slice());
-
-  validateBodyParameters(params, operationId);
-  validatePathParameters(params, pathId, operationId);
-  validateParameterTypes(params, api, operation, operationId);
 }
 
 /**
@@ -236,19 +185,53 @@ function validateParameterTypes(
 }
 
 /**
- * Checks the given parameter list for duplicates.
+ * Validates the parameters for the given operation.
  *
  */
-function checkForDuplicates(params: OpenAPIV2.ParameterObject[]) {
-  for (let i = 0; i < params.length - 1; i++) {
-    const outer = params[i];
-    for (let j = i + 1; j < params.length; j++) {
-      const inner = params[j];
-      if (outer.name === inner.name && outer.in === inner.in) {
-        throw ono.syntax(`Validation failed. Found multiple ${outer.in} parameters named "${outer.name}"`);
-      }
-    }
+function validateParameters(
+  api: OpenAPIV2.Document,
+  path: OpenAPIV2.PathItemObject,
+  pathId: string,
+  operation: OpenAPIV2.OperationObject,
+  operationId: string,
+) {
+  const pathParams = (path.parameters || []).filter(param => !('$ref' in param)) as OpenAPIV2.ParameterObject[];
+  const operationParams = (operation.parameters || []).filter(
+    param => !('$ref' in param),
+  ) as OpenAPIV2.ParameterObject[];
+
+  // Check for duplicate path parameters
+  try {
+    checkForDuplicates(pathParams);
+  } catch (e) {
+    throw ono.syntax(e, `Validation failed. ${pathId} has duplicate parameters`);
   }
+
+  // Check for duplicate operation parameters
+  try {
+    checkForDuplicates(operationParams);
+  } catch (e) {
+    throw ono.syntax(e, `Validation failed. ${operationId} has duplicate parameters`);
+  }
+
+  // Combine the path and operation parameters,
+  // with the operation params taking precedence over the path params
+  const params = pathParams.reduce((combinedParams, value) => {
+    const duplicate = combinedParams.some(param => {
+      if ('$ref' in param || '$ref' in value) {
+        return false;
+      }
+      return param.in === value.in && param.name === value.name;
+    });
+    if (!duplicate) {
+      combinedParams.push(value);
+    }
+    return combinedParams;
+  }, operationParams.slice());
+
+  validateBodyParameters(params, operationId);
+  validatePathParameters(params, pathId, operationId);
+  validateParameterTypes(params, api, operation, operationId);
 }
 
 /**
@@ -281,48 +264,65 @@ function validateResponse(code: number | string, response: OpenAPIV2.ResponseObj
 }
 
 /**
- * Validates the given Swagger schema object.
+ * Validates the given path.
  *
  */
-function validateSchema(schema: OpenAPIV2.SchemaObject, schemaId: string) {
-  if (schema.type === 'array' && !schema.items) {
-    throw ono.syntax(`Validation failed. ${schemaId} is an array, so it must include an "items" schema`);
-  }
+function validatePath(api: OpenAPIV2.Document, path: OpenAPIV2.PathItemObject, pathId: string, operationIds: string[]) {
+  swaggerHTTPMethods.forEach(operationName => {
+    const operation = path[operationName];
+    const operationId = `${pathId}/${operationName}`;
+
+    if (operation) {
+      const declaredOperationId = operation.operationId;
+      if (declaredOperationId) {
+        if (!operationIds.includes(declaredOperationId)) {
+          operationIds.push(declaredOperationId);
+        } else {
+          throw ono.syntax(`Validation failed. Duplicate operation id '${declaredOperationId}'`);
+        }
+      }
+      validateParameters(api, path, pathId, operation, operationId);
+
+      Object.keys(operation.responses || {}).forEach(responseName => {
+        const response = operation.responses[responseName];
+        if ('$ref' in response || !response) {
+          return;
+        }
+
+        const responseId = `${operationId}/responses/${responseName}`;
+        validateResponse(responseName, response, responseId);
+      });
+    }
+  });
 }
 
 /**
- * Validates that the declared properties of the given Swagger schema object actually exist.
+ * Validates parts of the Swagger 2.0 specification that aren't covered by its JSON Schema
+ * definition.
  *
+ * @see {@link https://github.com/OAI/OpenAPI-Specification/blob/main/versions/2.0.md}
  */
-function validateRequiredPropertiesExist(schema: IJsonSchema, schemaId: string) {
-  // Recursively collects all properties of the schema and its ancestors. They are added to the props object.
-  function collectProperties(schemaObj: IJsonSchema, props: Record<string, IJsonSchema>) {
-    if (schemaObj.properties) {
-      Object.keys(schemaObj.properties).forEach(property => {
-        // eslint-disable-next-line no-prototype-builtins
-        if (schemaObj.properties.hasOwnProperty(property)) {
-          // eslint-disable-next-line no-param-reassign
-          props[property] = schemaObj.properties[property];
-        }
-      });
+export function validateSpec(api: OpenAPIV2.Document) {
+  const operationIds: string[] = [];
+  Object.keys(api.paths || {}).forEach(pathName => {
+    const path = api.paths[pathName];
+    const pathId = `/paths${pathName}`;
+
+    if (path && pathName.startsWith('/')) {
+      validatePath(api, path, pathId, operationIds);
+    }
+  });
+
+  Object.keys(api.definitions || {}).forEach(definitionName => {
+    const definition = api.definitions[definitionName];
+    const definitionId = `/definitions/${definitionName}`;
+
+    if (!/^[a-zA-Z0-9.\-_]+$/.test(definitionName)) {
+      throw ono.syntax(
+        `Validation failed. ${definitionId} has an invalid name. Definition names should match against: /^[a-zA-Z0-9.-_]+$/`,
+      );
     }
 
-    if (schemaObj.allOf) {
-      schemaObj.allOf.forEach(parent => {
-        collectProperties(parent, props);
-      });
-    }
-  }
-
-  if (schema.required && Array.isArray(schema.required)) {
-    const props: Record<string, IJsonSchema> = {};
-    collectProperties(schema, props);
-    schema.required.forEach(requiredProperty => {
-      if (!props[requiredProperty]) {
-        throw ono.syntax(
-          `Validation failed. Property '${requiredProperty}' listed as required but does not exist in '${schemaId}'`,
-        );
-      }
-    });
-  }
+    validateRequiredPropertiesExist(definition, definitionId);
+  });
 }
